+++ internal/provider/compile_data_source_test.go
@@ -3,7 +3,7 @@ package provider
 import (
 	"context"
 	"fmt"
-	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
@@ -15,7 +15,6 @@ import (
 	"github.com/stevencyb/gopackager/internal/hasher"
 	"github.com/stevencyb/gopackager/internal/packager"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 )
 
 func TestAccDataSourceFrameworkSatisfaction(t *testing.T) {
@@ -25,21 +24,7 @@ func TestAccDataSourceFrameworkSatisfaction(t *testing.T) {
 }
 
 func TestAccCompileDataSource(t *testing.T) {
-	// Save original global instances
-	originalCompiler := globalCompiler
-	originalZIPPackager := globalZIPPackager
-	originalHasher := globalHasher
-
-	// Clean up after test
-	t.Cleanup(func() {
-		globalCompiler = originalCompiler
-		globalZIPPackager = originalZIPPackager
-		globalHasher = originalHasher
-		// Clean up any .gopackager metadata files created during the test
-		os.Remove("linux_amd64_binary.gopackager")
-		os.Remove("windows_amd64_binary.gopackager")
-		os.Remove("windows_amd64_binary.zip.gopackager")
-	})
+	t.Parallel()
 
 	mockCompiler := compiler.MockCompiler{}
 	mockPackager := packager.MockZIP{}
@@ -71,7 +56,6 @@ func TestAccCompileDataSource(t *testing.T) {
 		OutputSHA512:       types.StringValue("sha512hash"),
 		OutputSHA256Base64: types.StringValue("sha256base64hash"),
 		OutputSHA512Base64: types.StringValue("sha512base64hash"),
-		OutputGITHash:      types.StringValue("outputGITHash"),
 	}
 	firstUpdate := CompileDataSourceModel{
 		Source:             types.StringValue("provider.go"),
@@ -85,7 +69,6 @@ func TestAccCompileDataSource(t *testing.T) {
 		OutputSHA512:       types.StringValue("sha512hash"),
 		OutputSHA256Base64: types.StringValue("sha256base64hash"),
 		OutputSHA512Base64: types.StringValue("sha512base64hash"),
-		OutputGITHash:      types.StringValue("outputGITHash"),
 	}
 	secondUpdate := CompileDataSourceModel{
 		Source:             types.StringValue("provider.go"),
@@ -99,7 +82,6 @@ func TestAccCompileDataSource(t *testing.T) {
 		OutputSHA512:       types.StringValue("sha512hash"),
 		OutputSHA256Base64: types.StringValue("sha256base64hash"),
 		OutputSHA512Base64: types.StringValue("sha512base64hash"),
-		OutputGITHash:      types.StringValue("outputGITHash"),
 	}
 	thirdUpdate := CompileDataSourceModel{
 		Source:             types.StringValue("provider.go"),
@@ -113,13 +95,22 @@ func TestAccCompileDataSource(t *testing.T) {
 		OutputSHA512:       types.StringValue("sha512hash"),
 		OutputSHA256Base64: types.StringValue("sha256base64hash"),
 		OutputSHA512Base64: types.StringValue("sha512base64hash"),
-		OutputGITHash:      types.StringValue("outputGITHash"),
 		ZIP:                types.BoolValue(true),
 		ZIPResources:       additionalZIPResourcesGen,
 	}
 
-	mockHasher.On("ReadFile", initialDataSource.OutputPath.ValueString()).Return([]byte("123"), nil)
-	mockHasher.On("CombinedHash", []byte("123")).Return(hasher.CombinedHash{
+	mockHasher.On("ReadFile", initialDataSource.OutputPath.ValueString()).Times(3).Return([]byte("123"), nil)
+	mockHasher.On("CombinedHash", []byte("123")).Times(3).Return(hasher.CombinedHash{
+		MD5:          initialDataSource.OutputMD5.ValueString(),
+		SHA1:         initialDataSource.OutputSHA1.ValueString(),
+		SHA256:       initialDataSource.OutputSHA256.ValueString(),
+		SHA512:       initialDataSource.OutputSHA512.ValueString(),
+		SHA256Base64: initialDataSource.OutputSHA256Base64.ValueString(),
+		SHA512Base64: initialDataSource.OutputSHA512Base64.ValueString(),
+	}, nil)
+
+	basePath := filepath.Dir(initialDataSource.Source.ValueString())
+	mockHasher.On("HashDir", basePath).Return(&hasher.CombinedHash{
 		MD5:          initialDataSource.OutputMD5.ValueString(),
 		SHA1:         initialDataSource.OutputSHA1.ValueString(),
 		SHA256:       initialDataSource.OutputSHA256.ValueString(),
@@ -133,10 +124,11 @@ func TestAccCompileDataSource(t *testing.T) {
 			Destination(initialDataSource.Destination.ValueString()).
 			GOOS(initialDataSource.GOOS.ValueString()).
 			GOARCH(initialDataSource.GOARCH.ValueString()),
-	).Return(initialDataSource.OutputPath.ValueString(), nil)
+	).Times(3).
+		Return(initialDataSource.OutputPath.ValueString(), nil)
 
-	mockHasher.On("ReadFile", firstUpdate.OutputPath.ValueString()).Return([]byte("333"), nil)
-	mockHasher.On("CombinedHash", []byte("333")).Return(hasher.CombinedHash{
+	mockHasher.On("ReadFile", firstUpdate.OutputPath.ValueString()).Times(6).Return([]byte("333"), nil)
+	mockHasher.On("CombinedHash", []byte("333")).Times(6).Return(hasher.CombinedHash{
 		MD5:          firstUpdate.OutputMD5.ValueString(),
 		SHA1:         firstUpdate.OutputSHA1.ValueString(),
 		SHA256:       firstUpdate.OutputSHA256.ValueString(),
@@ -150,26 +142,28 @@ func TestAccCompileDataSource(t *testing.T) {
 			Destination(firstUpdate.Destination.ValueString()).
 			GOOS(firstUpdate.GOOS.ValueString()).
 			GOARCH(firstUpdate.GOARCH.ValueString()),
-	).Return(firstUpdate.OutputPath.ValueString(), nil)
+	).Times(3).
+		Return(firstUpdate.OutputPath.ValueString(), nil)
 
-	// ReadFile and CombinedHash are reused for secondUpdate
+	// ReadFile and CombinedHash are reused
 	mockCompiler.On("Compile",
 		*compiler.NewConfig().
 			Source(secondUpdate.Source.ValueString()).
 			Destination(secondUpdate.Destination.ValueString()).
 			GOOS(secondUpdate.GOOS.ValueString()).
 			GOARCH(secondUpdate.GOARCH.ValueString()),
-	).Return(secondUpdate.OutputPath.ValueString(), nil)
+	).Times(3).
+		Return(secondUpdate.OutputPath.ValueString(), nil)
 
-	mockPackager.On("Zip", thirdUpdate.OutputPath.ValueString()+".zip", additionalZIPResources).Return(nil)
-	mockHasher.On("ReadFile", thirdUpdate.OutputPath.ValueString()+".zip").Return([]byte("666"), nil)
-	mockHasher.On("CombinedHash", []byte("666")).Return(hasher.CombinedHash{
-		MD5:          thirdUpdate.OutputMD5.ValueString(),
-		SHA1:         thirdUpdate.OutputSHA1.ValueString(),
-		SHA256:       thirdUpdate.OutputSHA256.ValueString(),
-		SHA512:       thirdUpdate.OutputSHA512.ValueString(),
-		SHA256Base64: thirdUpdate.OutputSHA256Base64.ValueString(),
-		SHA512Base64: thirdUpdate.OutputSHA512Base64.ValueString(),
+	mockPackager.On("Zip", thirdUpdate.OutputPath.ValueString()+".zip", additionalZIPResources).Times(3).Return(nil)
+	mockHasher.On("ReadFile", thirdUpdate.OutputPath.ValueString()+".zip").Times(3).Return([]byte("666"), nil)
+	mockHasher.On("CombinedHash", []byte("666")).Times(3).Return(hasher.CombinedHash{
+		MD5:          firstUpdate.OutputMD5.ValueString(),
+		SHA1:         firstUpdate.OutputSHA1.ValueString(),
+		SHA256:       firstUpdate.OutputSHA256.ValueString(),
+		SHA512:       firstUpdate.OutputSHA512.ValueString(),
+		SHA256Base64: firstUpdate.OutputSHA256Base64.ValueString(),
+		SHA512Base64: firstUpdate.OutputSHA512Base64.ValueString(),
 	}, nil)
 	mockCompiler.On("Compile",
 		*compiler.NewConfig().
@@ -177,7 +171,8 @@ func TestAccCompileDataSource(t *testing.T) {
 			Destination(thirdUpdate.Destination.ValueString()).
 			GOOS(thirdUpdate.GOOS.ValueString()).
 			GOARCH(thirdUpdate.GOARCH.ValueString()),
-	).Return(thirdUpdate.OutputPath.ValueString(), nil)
+	).Times(3).
+		Return(thirdUpdate.OutputPath.ValueString(), nil)
 
 	resource.Test(t, resource.TestCase{
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -257,93 +252,14 @@ func TestAccCompileDataSource(t *testing.T) {
 	})
 }
 
-func TestAccCompileDataSourceWithGitTrigger(t *testing.T) {
-	// Save original global instances
-	originalCompiler := globalCompiler
-	originalZIPPackager := globalZIPPackager
-	originalHasher := globalHasher
-
-	// Clean up after test
-	t.Cleanup(func() {
-		globalCompiler = originalCompiler
-		globalZIPPackager = originalZIPPackager
-		globalHasher = originalHasher
-	})
-
-	mockCompiler := compiler.MockCompiler{}
-	mockPackager := packager.MockZIP{}
-	mockHasher := hasher.MockHasher{}
-	globalCompiler = &mockCompiler
-	globalZIPPackager = &mockPackager
-	globalHasher = &mockHasher
-
-	// Test data source with git trigger enabled
-	gitTriggerDataSource := CompileDataSourceModel{
-		Source:             types.StringValue("provider.go"),
-		Destination:        types.StringValue("linux_amd64_binary"),
-		GOOS:               types.StringValue("linux"),
-		GOARCH:             types.StringValue("amd64"),
-		GitTrigger:         types.BoolValue(true),
-		GitTriggerPath:     types.StringValue("./src"),
-		OutputPath:         types.StringValue("linux_amd64_binary"),
-		OutputMD5:          types.StringValue("md5hash"),
-		OutputSHA1:         types.StringValue("sha1hash"),
-		OutputSHA256:       types.StringValue("sha256hash"),
-		OutputSHA512:       types.StringValue("sha512hash"),
-		OutputSHA256Base64: types.StringValue("sha256base64hash"),
-		OutputSHA512Base64: types.StringValue("sha512base64hash"),
-		OutputGITHash:      types.StringValue("outputGITHash"),
-	}
-
-	mockHasher.On("ReadFile", gitTriggerDataSource.OutputPath.ValueString()).Return([]byte("123"), nil)
-	mockHasher.On("CombinedHash", []byte("123")).Return(hasher.CombinedHash{
-		MD5:          "md5hash",
-		SHA1:         "sha1hash",
-		SHA256:       "sha256hash",
-		SHA512:       "sha512hash",
-		SHA256Base64: "sha256base64hash",
-		SHA512Base64: "sha512base64hash",
-	})
-	mockCompiler.On("Compile", mock.AnythingOfType("compiler.Config")).Return("linux_amd64_binary", nil)
-
-	testAccProtoV6ProviderFactories := map[string]func() (tfprotov6.ProviderServer, error){
-		"gopackager": providerserver.NewProtocol6WithError(New("test")()),
-	}
-
-	resource.Test(t, resource.TestCase{
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: compilerDataSourceFromModel(t, gitTriggerDataSource),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("data.gopackager_compile.test", "source", "provider.go"),
-					resource.TestCheckResourceAttr("data.gopackager_compile.test", "git_trigger", "true"),
-					resource.TestCheckResourceAttr("data.gopackager_compile.test", "git_trigger_path", "./src"),
-					resource.TestCheckResourceAttr("data.gopackager_compile.test", "output_path", "linux_amd64_binary"),
-				),
-			},
-		},
-	})
-}
-
 func compilerDataSourceFromModel(t *testing.T, model CompileDataSourceModel) string {
 	t.Helper()
 
 	zip := ""
 	zipResource := ""
-	gitTrigger := ""
-	gitTriggerPath := ""
 
 	if !model.ZIP.IsNull() && !model.ZIP.IsUnknown() && model.ZIP.ValueBool() {
 		zip = `zip = true`
-	}
-
-	if !model.GitTrigger.IsNull() && !model.GitTrigger.IsUnknown() && model.GitTrigger.ValueBool() {
-		gitTrigger = `git_trigger = true`
-	}
-
-	if !model.GitTriggerPath.IsNull() && !model.GitTriggerPath.IsUnknown() {
-		gitTriggerPath = fmt.Sprintf(`git_trigger_path = %s`, model.GitTriggerPath.String())
 	}
 
 	if !model.ZIPResources.IsNull() && !model.ZIPResources.IsUnknown() {
@@ -367,8 +283,6 @@ data "gopackager_compile" "test" {
 	goarch = %s
 	%s
 	%s
-	%s
-	%s
 }
-	`, model.Source.String(), model.Destination.String(), model.GOOS.String(), model.GOARCH.String(), zip, zipResource, gitTrigger, gitTriggerPath)
+	`, model.Source.String(), model.Destination.String(), model.GOOS.String(), model.GOARCH.String(), zip, zipResource)
 }